move RuleConfiguration
